requirements.txt
# Azure
kafka-python
pydantic
pydantic-settings
psycopg2-binary
